// src/app/model/productos_model.ts
  
  export class ProductModel{
    prodId: string='';
    prodDesc: string = '';
    prodCode: string = '';
    prodPrice: string='';
    prodStock: number=0;
    prodState: boolean = false;
    catId: number=0;
    catName: string = '';
    supplierDetailId: number=0;
    supDetId: number=0;
    supDetFactory: string = "";
    supDetAddress: string = "";
    supDetNumber: string = "";
    supplierId: number=0;
    supplierRazonSocial: string = '';
    supplierRuc: number=0;
    supplierContacto: number=0;
    typeId: number=0;
    typeName: string='';
    locationId: number=0;
    locationFloor: string='';
  }